fit login module into database setup
