+++ .gitignore
@@ -9,3 +9,4 @@
 src.egg-info
 venv/
 .env
+__pycache__
