+++ requirements.txt
@@ -1,14 +1,26 @@
+attrs==19.3.0
 click==7.1.2
+flake8==3.8.1
 Flask==1.1.2
+Flask-JWT-Extended==3.24.1
 gunicorn==20.0.4
+importlib-metadata==1.6.0
 itsdangerous==1.1.0
 Jinja2==2.11.2
 MarkupSafe==1.1.1
+mccabe==0.6.1
+more-itertools==8.3.0
+packaging==20.3
+pluggy==0.13.1
 psycopg2-binary==2.8.5
+py==1.8.1
+pycodestyle==2.6.0
+pyflakes==2.2.0
+PyJWT==1.7.1
+pyparsing==2.4.7
+pytest==5.4.2
 python-decouple==3.3
-flake8==3.7.9
-Flask-JWT-Extended==3.24.1
-pycodestyle==2.5.0
-pyflakes==2.1.1
-pytest==5.4.1
+six==1.14.0
+wcwidth==0.1.9
 Werkzeug==1.0.1
+zipp==3.1.0
